+++ zellij-utils/src/input/options.rs
@@ -52,6 +52,9 @@ pub struct Options {
     /// Set the default shell
     #[clap(long, parse(from_os_str))]
     pub default_shell: Option<PathBuf>,
+    /// Set the default layout
+    #[clap(long, parse(from_os_str))]
+    pub default_layout: Option<PathBuf>,
     /// Set the layout_dir, defaults to
     /// subdirectory of config dir
     #[clap(long, parse(from_os_str))]
@@ -89,6 +92,10 @@ pub struct Options {
     #[clap(long)]
     #[serde(default)]
     pub copy_on_select: Option<bool>,
+
+    /// Explicit full path to open the scrollback editor (default is $EDITOR or $VISUAL)
+    #[clap(long, parse(from_os_str))]
+    pub scrollback_editor: Option<PathBuf>,
 }
 
 #[derive(ArgEnum, Deserialize, Serialize, Debug, Clone, Copy, PartialEq)]
@@ -124,6 +131,7 @@ impl Options {
         let simplified_ui = other.simplified_ui.or(self.simplified_ui);
         let default_mode = other.default_mode.or(self.default_mode);
         let default_shell = other.default_shell.or_else(|| self.default_shell.clone());
+        let default_layout = other.default_layout.or_else(|| self.default_layout.clone());
         let layout_dir = other.layout_dir.or_else(|| self.layout_dir.clone());
         let theme = other.theme.or_else(|| self.theme.clone());
         let on_force_close = other.on_force_close.or(self.on_force_close);
@@ -131,12 +139,16 @@ impl Options {
         let copy_command = other.copy_command.or_else(|| self.copy_command.clone());
         let copy_clipboard = other.copy_clipboard.or(self.copy_clipboard);
         let copy_on_select = other.copy_on_select.or(self.copy_on_select);
+        let scrollback_editor = other
+            .scrollback_editor
+            .or_else(|| self.scrollback_editor.clone());
 
         Options {
             simplified_ui,
             theme,
             default_mode,
             default_shell,
+            default_layout,
             layout_dir,
             mouse_mode,
             pane_frames,
@@ -146,6 +158,7 @@ impl Options {
             copy_command,
             copy_clipboard,
             copy_on_select,
+            scrollback_editor,
         }
     }
 
@@ -171,6 +184,7 @@ impl Options {
 
         let default_mode = other.default_mode.or(self.default_mode);
         let default_shell = other.default_shell.or_else(|| self.default_shell.clone());
+        let default_layout = other.default_layout.or_else(|| self.default_layout.clone());
         let layout_dir = other.layout_dir.or_else(|| self.layout_dir.clone());
         let theme = other.theme.or_else(|| self.theme.clone());
         let on_force_close = other.on_force_close.or(self.on_force_close);
@@ -178,12 +192,16 @@ impl Options {
         let copy_command = other.copy_command.or_else(|| self.copy_command.clone());
         let copy_clipboard = other.copy_clipboard.or(self.copy_clipboard);
         let copy_on_select = other.copy_on_select.or(self.copy_on_select);
+        let scrollback_editor = other
+            .scrollback_editor
+            .or_else(|| self.scrollback_editor.clone());
 
         Options {
             simplified_ui,
             theme,
             default_mode,
             default_shell,
+            default_layout,
             layout_dir,
             mouse_mode,
             pane_frames,
@@ -193,6 +211,7 @@ impl Options {
             copy_command,
             copy_clipboard,
             copy_on_select,
+            scrollback_editor,
         }
     }
 
@@ -235,6 +254,7 @@ impl From<CliOptions> for Options {
             theme: opts.theme,
             default_mode: opts.default_mode,
             default_shell: opts.default_shell,
+            default_layout: opts.default_layout,
             layout_dir: opts.layout_dir,
             mouse_mode: opts.mouse_mode,
             pane_frames: opts.pane_frames,
@@ -244,6 +264,7 @@ impl From<CliOptions> for Options {
             copy_command: opts.copy_command,
             copy_clipboard: opts.copy_clipboard,
             copy_on_select: opts.copy_on_select,
+            scrollback_editor: opts.scrollback_editor,
         }
     }
 }